v9.99.12 — Chạy demo tấm TẦM NHÌN, và hai chỗ bộ kiểm không thấy
Anh bảo cho app chạy thử để thiết kế ảnh cho tầm nhìn GITA365. Tôi lấy
đúng chữ TẦM NHÌN · MỐC 2030 · SỨ MỆNH trong G.CULTURE — bản gốc DUY
NHẤT của ba câu ấy — đẩy qua cổng Kiến trúc sư thị giác. Cổng làm đúng
việc, nhưng đường đi lộ ra hai chỗ mà 989 phép đo không thấy.

CỔNG LÀM ĐÚNG, ĐÃ ĐO
  · Tấm bìa tầm nhìn ở chặng T1 QUA cổng Tầng. Chữ tầm nhìn thật không
    hứa gì ngoài phạm vi T1 — nó nói về loại thế giới muốn tạo ra, không
    nói về thứ chặng nào giao.
  · Cùng chữ ấy mà gắn thêm "coach đồng hành · phác đồ · kho nghề" thì
    cổng CHẶN, và nói ra đủ ba chỗ phạm. Đây đúng là chỗ tấm tầm nhìn
    dễ hỏng nhất: câu tầm nhìn nói cả đời người, người làm hình rất dễ
    kéo lời hứa chặng cao xuống tấm mở đầu.
  · Nhảy tắt deXuat → duyet: chặn, mã SAIBAC. Không có đường tắt.
  · Cửa đi ra: ĐÓNG, vì chưa nạp GITA_KHOA_VE và GITA_CONG_VE.

CHỖ THỨ NHẤT — SỔ TỰ CÃI MÌNH
Chấm tấm tầm nhìn ra 89,9 điểm, bậc "Sửa lại". Sổ ghi xuống: 90.
Mà 90 đúng là ngưỡng của bậc "Đạt".

Người đọc sổ sáu tháng sau thấy "90đ · Sửa lại" và không biết tin cái
nào. Nguyên nhân: bậc tính trên số LẺ, còn cột diem khai INTEGER nên
chỗ ghi phải Math.round. Mọi bài thử cũ đều chấm ra số CHẴN — 75 và 90
— nên chỗ này im suốt từ lúc dựng.

Chữa bằng cách giữ SỐ THẬT: cột thành REAL, bỏ Math.round. Không chọn
cách ngược lại (làm tròn trước rồi mới xếp bậc) vì làm tròn LÊN là đẩy
một bài trượt qua cửa đạt, mà cửa ấy là thứ cổng này dựng ra để canh.

CHỖ THỨ HAI — MỘT CHỮ "VÀ" ĐI LỌT MÀ KHÔNG AI BIẾT
"Nói tầm nhìn và giới thiệu năm chặng đồng hành" — hai việc rõ ràng,
đi lọt sạch, không một dòng cảnh báo.

Ngưỡng chặn đặt ở HAI chữ "và" là cố ý và vẫn giữ: "cho phụ huynh và
học viên" là một việc, chặn nó là chặn oan. Máy không phân biệt được
hai câu ấy — nên máy KHÔNG quyết. Nó nói ra chỗ đáng ngờ, và câu ấy
đi theo đề bài tới tận bậc duyệt, hiện cả trên màn lúc vừa đề xuất.
Đúng luật của cổng này: máy đề xuất, chủ hệ quyết.

BA PHÉP KIỂM MỚI, ĐÃ PHÁ THỬ TỪNG CÁI
  · trả Math.round lại  → đỏ đúng phép "số lẻ ghi nguyên số lẻ"
  · bỏ lưu ý            → đỏ đúng phép "một chữ và không im"
  · dán lưu ý mọi tấm   → đỏ đúng phép "không có và thì không dán"
Cảnh báo dán vào mọi tấm là cảnh báo không ai đọc, nên phép thứ ba
canh chiều ngược lại chứ không chỉ canh chiều thiếu.

989 phép đo · trợ lý 28/40 trúng · 32/40 có mặt · rà soát sạch
880 kho · 0 đổi · 0 biến mất

+++ may-chu/csdl.sql
@@ -740,7 +740,13 @@ CREATE TABLE IF NOT EXISTS deXuatThiGiac (
   -- lại: ranh giới Tầng đổi thì đề xuất cũ vẫn phải giải thích được là
   -- nó đã qua cổng nào.
   soatTang    TEXT,
-  diem        INTEGER,
+  -- REAL, không INTEGER. Trọng số là phần trăm nên tổng gần như luôn lẻ:
+  -- 89,9 chẳng hạn. Bản đầu để INTEGER và ghi Math.round(tổng), trong khi
+  -- BẬC lại tính trên số LẺ — nên sổ hiện "90đ · Sửa lại", mà 90 chính là
+  -- ngưỡng của bậc ĐẠT. Người đọc sổ sáu tháng sau thấy hai thứ cãi nhau.
+  -- Giữ số thật thì bậc và số luôn cùng một gốc, và không có bài nào bị
+  -- làm tròn LÊN qua ngưỡng đạt.
+  diem        REAL,
   bacDiem     TEXT,
   chamChiTiet TEXT,               -- JSON điểm từng mục
   trangThai   TEXT NOT NULL DEFAULT 'nhap',
